+++ src/faster/api/cvkvcontext.h
@@ -315,7 +315,7 @@ protected:
 private:
     Key key_;
 public:
-    uint8_t output_length;
+    uint32_t output_length;
     // Extract two bytes of output.
     uint8_t *output_bytes;
 };
